numpy 2 compatible (#249)
* These changes enable calcos to run under numpy 2, and also lower the differences
between running with numpy 2 and numpy 1.x to more tolerable levels

* Remove npy_promotion_state calls and fix an indent

* Changes for numpy 2:
 - casting changes required for C extensions
 - casting numpy float32 scalars to float64 to ensure that arithmetic is done using float64 precision
 - remove some conditionals in the C extension that were put in for python 2/3 compatibility

* Don't cast arrays that are None

+++ calcos/dispersion.py
@@ -134,7 +134,7 @@ class Dispersion(object):
             Wavelength (or array of wavelengths) at x
         """
 
-        x_prime = x + self.delta
+        x_prime = np.float64(x) + np.float64(self.delta)
 
         sum = self.coeff[self.ncoeff-1]
         for i in range(self.ncoeff-2, -1, -1):
@@ -159,7 +159,7 @@ class Dispersion(object):
             Slope at x, in Angstroms per pixel
         """
 
-        x_prime = x + self.delta
+        x_prime = np.float64(x) + np.float64(self.delta)
 
         sum = (self.ncoeff - 1.) * self.coeff[self.ncoeff-1]
         for n in range(self.ncoeff-2, 0, -1):
